backend/requirements.txt
fastapi
uvicorn[standard]
httpx
google-genai
pydantic
pydantic-settings
python-dotenv
